Login route returns member ID and sets sessions cookie
The /login route now returns a JSON object with the member's ID (Primary Key).  This should be saved in the local store by the frontend to determine other data.
