+++ authentications/sessions.go
@@ -1,10 +1,15 @@
 package gatekeeper
 
 import (
+	"encoding/json"
 	"fmt"
 	"github.com/gorilla/sessions"
 	"net/http"
 )
+
+type memberDetails struct {
+	ID string
+}
 
 var (
 	// key must be 16, 24 or 32 bytes long (AES-128, AES-192 or AES-256)
@@ -18,14 +23,11 @@ func ValidSession(w http.ResponseWriter, r *http.Request) {
 	if err != nil {
 		panic(err)
 	}
-
 	// Check if user is authenticated
 	if auth, ok := session.Values["authenticated"].(bool); !ok || !auth {
 		http.Error(w, "Is this session valid: false", http.StatusUnauthorized)
 		return
 	}
-
-	// Return message
 	fmt.Fprintln(w, "Is this session valid: true")
 }
 
@@ -33,27 +35,27 @@ func ValidSession(w http.ResponseWriter, r *http.Request) {
 func Login(w http.ResponseWriter, r *http.Request) {
 	session, err := store.Get(r, "scheduler-session")
 	if err != nil {
-		panic(err)
+		fmt.Println(err)
 	}
+	creds := DecodeCredentials(r)
 	// Authenticate based on incoming http request
-	if passwordsMatch(r) != true {
+	if passwordsMatch(r, creds) != true {
 		http.Error(w, "Incorrect username or password", http.StatusUnauthorized)
 		return
 	}
-	
-	// Assume passwords matched, add a function to get the memberID based ont the email
-	fmt.Println(getMemberID("Jack_Reinger@yahoo.com"))
-	
+	// Get the memberID based on the supplied email
+	memberID := getMemberID(creds.Email)
+	m := memberDetails {
+		ID: memberID,
+	}
+
 	// Respond with the proper content type and the memberID
-	w.Header().Set("Content-Type", "text/plain") // TODO convert this to application/json
-	w.WriteHeader(http.StatusOK)
-	memberID := "1234"
-	fmt.Fprintf(w, memberID)
-	// w.Write() // Alternative to fprintf.  Needs []byte of marshalled JSON
-	
+	w.Header().Set("Content-Type", "application/json") // TODO convert this to application/json
 	// Set cookie values and save
 	session.Values["authenticated"] = true
 	session.Save(r, w)
+	json.NewEncoder(w).Encode(m)
+	// w.Write([]byte(memberID)) // Alternative to fprintf
 }
 
 // Logout destroys the session
@@ -62,7 +64,6 @@ func Logout(w http.ResponseWriter, r *http.Request) {
 	if err != nil {
 		panic(err)
 	}
-
 	// Revoke users authentication
 	session.Values["authenticated"] = false
 	session.Save(r, w)
